+++ contracts/access/MinterRole.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.6.7;
+pragma solidity ^0.6.8;
 
 import "@openzeppelin/contracts/access/AccessControl.sol";
 
@@ -18,18 +18,18 @@ contract MinterRole is AccessControl {
     }
 
     function isMinter(address account) public view returns (bool) {
-        require(account != address(0));
+        require(account != address(0), "MinterRole: address zero cannot be minter");
         return hasRole(DEFAULT_ADMIN_ROLE, account);
     }
 
     function addMinter(address account) public onlyMinter {
-        require(!isMinter(account));
+        require(!isMinter(account), "MinterRole: add an account already minter");
         grantRole(DEFAULT_ADMIN_ROLE, account);
         emit MinterAdded(account);
     }
 
     function renounceMinter() public {
-        require(isMinter(_msgSender()));
+        require(isMinter(_msgSender()), "MinterRole: renounce by a non-minter");
         renounceRole(DEFAULT_ADMIN_ROLE, _msgSender());
         emit MinterRemoved(_msgSender());
     }