+++ contracts/access/PauserRole.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.6.7;
+pragma solidity ^0.6.8;
 
 import "@openzeppelin/contracts/access/AccessControl.sol";
 
@@ -18,18 +18,18 @@ contract PauserRole is AccessControl {
     }
 
     function isPauser(address account) public view returns (bool) {
-        require(account != address(0));
+        require(account != address(0), "PauserRole: address zero cannot be pauser");
         return hasRole(DEFAULT_ADMIN_ROLE, account);
     }
 
     function addPauser(address account) public onlyPauser {
-        require(!isPauser(account));
+        require(!isPauser(account), "PauserRole: add an account already pauser");
         grantRole(DEFAULT_ADMIN_ROLE, account);
         emit PauserAdded(account);
     }
 
     function renouncePauser() public {
-        require(isPauser(_msgSender()));
+        require(isPauser(_msgSender()), "PauserRole: renounce by a non-pauser");
         renounceRole(DEFAULT_ADMIN_ROLE, _msgSender());
         emit PauserRemoved(_msgSender());
     }