+++ contracts/algo/EnumSet.sol
@@ -25,7 +25,7 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-pragma solidity ^0.6.7;
+pragma solidity ^0.6.8;
 
 /**
  * @dev Library for managing
@@ -57,11 +57,11 @@ library EnumSet {
 
     struct Set {
         // Storage of set values
-        bytes32[] _values;
+        bytes32[] values;
 
         // Position of the value in the `values` array, plus 1 because index 0
         // means a value is not in the set.
-        mapping (bytes32 => uint256) _indexes;
+        mapping (bytes32 => uint256) indexes;
     }
 
     /**
@@ -72,10 +72,10 @@ library EnumSet {
      */
     function add(Set storage set, bytes32 value) internal returns (bool) {
         if (!contains(set, value)) {
-            set._values.push(value);
+            set.values.push(value);
             // The value is stored at length-1, but we add 1 to all indexes
             // and use 0 as a sentinel value
-            set._indexes[value] = set._values.length;
+            set.indexes[value] = set.values.length;
             return true;
         } else {
             return false;
@@ -90,31 +90,31 @@ library EnumSet {
      */
     function remove(Set storage set, bytes32 value) internal returns (bool) {
         // We read and store the value's index to prevent multiple reads from the same storage slot
-        uint256 valueIndex = set._indexes[value];
+        uint256 valueIndex = set.indexes[value];
 
         if (valueIndex != 0) { // Equivalent to contains(set, value)
-            // To delete an element from the _values array in O(1), we swap the element to delete with the last one in
+            // To delete an element from the values array in O(1), we swap the element to delete with the last one in
             // the array, and then remove the last element (sometimes called as 'swap and pop').
             // This modifies the order of the array, as noted in {at}.
 
             uint256 toDeleteIndex = valueIndex - 1;
-            uint256 lastIndex = set._values.length - 1;
+            uint256 lastIndex = set.values.length - 1;
 
             // When the value to delete is the last one, the swap operation is unnecessary. However, since this occurs
             // so rarely, we still do the swap anyway to avoid the gas cost of adding an 'if' statement.
 
-            bytes32 lastvalue = set._values[lastIndex];
+            bytes32 lastvalue = set.values[lastIndex];
 
             // Move the last value to the index where the value to delete is
-            set._values[toDeleteIndex] = lastvalue;
+            set.values[toDeleteIndex] = lastvalue;
             // Update the index for the moved value
-            set._indexes[lastvalue] = toDeleteIndex + 1; // All indexes are 1-based
+            set.indexes[lastvalue] = toDeleteIndex + 1; // All indexes are 1-based
 
             // Delete the slot where the moved value was stored
-            set._values.pop();
+            set.values.pop();
 
             // Delete the index for the deleted slot
-            delete set._indexes[value];
+            delete set.indexes[value];
 
             return true;
         } else {
@@ -126,14 +126,14 @@ library EnumSet {
      * @dev Returns true if the value is in the set. O(1).
      */
     function contains(Set storage set, bytes32 value) internal view returns (bool) {
-        return set._indexes[value] != 0;
+        return set.indexes[value] != 0;
     }
 
     /**
      * @dev Returns the number of values on the set. O(1).
      */
     function length(Set storage set) internal view returns (uint256) {
-        return set._values.length;
+        return set.values.length;
     }
 
    /**
@@ -147,7 +147,7 @@ library EnumSet {
     * - `index` must be strictly less than {length}.
     */
     function at(Set storage set, uint256 index) internal view returns (bytes32) {
-        require(set._values.length > index, "EnumSet: index out of bounds");
-        return set._values[index];
+        require(set.values.length > index, "EnumSet: index out of bounds");
+        return set.values[index];
     }
 }