+++ contracts/mocks/utils/types/UInt256BitsMock.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.6.7;
+pragma solidity ^0.6.8;
 
 import "../../../utils/types/UInt256Extract.sol";
 import "../../../utils/types/UInt256Inject.sol";