+++ contracts/mocks/utils/types/UInt256ToStringMock.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.6.7;
+pragma solidity ^0.6.8;
 
 import "../../../utils/types/UInt256ToDecimalString.sol";
 import "../../../utils/types/UInt256ToHexString.sol";