+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Breaking changes
  * Reorganised the contracts in `utils/types`, following the logic of one contract per conversion function to have fine control on amount of code in the compilation unit. The conversion libraries now only have internal functions, have mocks and some have unit tests.
  * Removed `BitsLayoutDecoder`.
- * Migrated compiler to `solc:0.6.7`.
+ * Migrated compiler to `solc:0.6.8`.
 
 ### Improvements
  * `MinterRole.sol` public functions will revert if supplied a zero-address target.